requirements.txt
# Bibliotecas de Visão Computacional e IA
opencv-python-headless
deepface
tf-keras
tensorflow==2.15.0

# Integração com AWS e Ambiente
boto3
python-dotenv

numpy<2.0.0